+++ include/generic.h
@@ -1,12 +1,13 @@
+#include <stdint.h>
 #ifndef _GENERIC_METHODS_H_
 #define _GENERIC_METHODS_H_
 
 typedef void (* GenFunction_0)(void);
-typedef void (* GenFunction_1)(s32);
-typedef void (* GenFunction_2)(s32, s32);
-typedef void (* GenFunction_3)(s32, s32, s32);
-typedef void (* GenFunction_4)(s32, s32, s32, s32);
-typedef void (* GenFunction_5)(s32, s32, s32, s32, s32);
+typedef void (* GenFunction_1)(intptr_t);
+typedef void (* GenFunction_2)(intptr_t, intptr_t);
+typedef void (* GenFunction_3)(intptr_t, intptr_t, intptr_t);
+typedef void (* GenFunction_4)(intptr_t, intptr_t, intptr_t, intptr_t);
+typedef void (* GenFunction_5)(intptr_t, intptr_t, intptr_t, intptr_t, intptr_t);
 typedef void (* GenFunction_6)(void *);
 
 #endif
